Move general factory packet processors
