+++ NebulaNetwork/PacketProcessors/Factory/CreatePrebuildsRequestProcessor.cs
@@ -8,7 +8,7 @@ using NebulaWorld;
 
 #endregion
 
-namespace NebulaNetwork.PacketProcessors.Factory.Entity;
+namespace NebulaNetwork.PacketProcessors.Factory;
 
 [RegisterPacketProcessor]
 internal class CreatePrebuildsRequestProcessor : PacketProcessor<CreatePrebuildsRequest>
