+++ NebulaNetwork/PacketProcessors/Factory/EntityBoostSwitchProcessor.cs
@@ -8,7 +8,7 @@ using NebulaModel.Packets.Factory;
 
 #endregion
 
-namespace NebulaNetwork.PacketProcessors.Factory.Entity;
+namespace NebulaNetwork.PacketProcessors.Factory;
 
 [RegisterPacketProcessor]
 internal class EntityBoostSwitchProcessor : PacketProcessor<EntityBoostSwitchPacket>
