+++ NebulaNetwork/PacketProcessors/Factory/ExtraInfoUpdateProcessor.cs
@@ -8,7 +8,7 @@ using NebulaWorld;
 
 #endregion
 
-namespace NebulaNetwork.PacketProcessors.Factory.Entity;
+namespace NebulaNetwork.PacketProcessors.Factory;
 
 [RegisterPacketProcessor]
 internal class ExtraInfoUpdateProcessor : PacketProcessor<ExtraInfoUpdatePacket>
